requirements.txt
auto-face-recognition==0.0.2
cloudpickle==1.6.0
cmake==3.21.2
cvlib==0.2.6
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.5.2
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
heroku==0.1.4
numpy==1.19.5
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
pickleshare==0.7.5
Pillow==8.2.0
python-dateutil==1.5
torch==1.10.0
torchvision==0.11.1
